refactor: dnd여부에 따라 PickListViewer분리
